+++ addons/Dexie.Syncable/src/globals.js
@@ -1,0 +1,3 @@
+export default {
+  db: null
+};
